fix: auth login flow, schema migrations, and UI updates
- Add login content component with auth types for extensible login flows
- Add auth session API route
- Add pre-push hook for CI checks
- Add input-otp UI component
- Add Resend bad emails and webhook migrations
- Pluralize table names migration
- Update Prisma schema with latest model changes
- Update turbo.json and package configs
