+++ .husky/pre-push
@@ -1,0 +1,1 @@
+pnpm run check
