+++ apps/web/package.json
@@ -18,6 +18,7 @@
     "@onecli/db": "workspace:*",
     "@onecli/ui": "workspace:*",
     "aws-amplify": "^6.16.2",
+    "input-otp": "^1.4.2",
     "lucide-react": "^0.562.0",
     "next": "16.1.0",
     "next-auth": "5.0.0-beta.30",
